fix: use stable unique ID in fn_cache key to prevent ABA stale hits
Rc::as_ptr was used as the function identity for memoization. When a
function's Rc dropped and a new UserFunction was allocated at the same
address, the cache returned a stale result — causing the nonlocal test
to non-deterministically print 2 instead of [5, 5] (inner()'s cached
return aliased outer_nonlocal's new Rc address).

Fix: add a monotonically increasing u64 id to UserFunction (via an
AtomicU64 counter in pyrust-core) and use function.id as the cache key
instead of the Rc pointer. The key type changes from (usize, Vec<PyKey>)
to (u64, Vec<PyKey>) in both the variadic call path and the inline
CallMemo path.

Fixes #87.

+++ src/interpreter.rs
@@ -14,7 +14,7 @@ use crate::value::{
 };
 
 type ModuleCache = Rc<RefCell<HashMap<String, Value>>>;
-type FnCache = HashMap<(usize, Vec<PyKey>), Value>;
+type FnCache = HashMap<(u64, Vec<PyKey>), Value>;
 
 const MAX_CALL_DEPTH: usize = 1000;
 const ENV_POOL_MAX: usize = 64;
